LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/installing.https-expected.txt


PASS installing is set
PASS The ServiceWorker objects returned from installing attribute getter that represent the same service worker are the same objects
